requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
h11==0.14.0
idna==3.10
iniconfig==2.0.0
outcome==1.3.0.post0
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
#PyQt6==6.8.1
#PyQt6-Qt6==6.8.2
#PyQt6_sip==13.10.0
PySocks==1.7.1
pytest==8.3.4
requests==2.32.3
selenium==4.27.1
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.3.0
websocket-client==1.8.0
webdriver-manager==4.0.2
wsproto==1.2.0
